refactor: extract code generation templates from ModelResourceJavaGenerator
Extract 6 template methods from buildJavaCodeBody into new package-private
ResourceCodeTemplates class:
- appendPreambleAndEnumConstants
- appendJointDeclarations (returns rootJoints list)
- appendRootJointIds
- appendPoseFields
- appendArrayFields
- appendConstructorsAndMethods

ModelResourceJavaGenerator reduced from 607 to 403 lines (under 500 target).
ResourceCodeTemplates is 324 lines with all static, stateless methods.

Fix 3 TDD test assumptions:
- Use FINGER_00 + exposeFirstArrays for array hint test (index must be 0)
- Use QuadrupedResource instead of BipedResource for roots tests
  (BipedResource has no methods returning JointId[])

All 230 tests pass including 41 new ResourceCodeTemplates tests.
